requirements.txt
numpy~=1.24.3
scikit-learn~=1.3.0
tensorflow==2.13.0
keras~=2.13.1
Pillow~=9.5.0
matplotlib~=3.7.1
scikit-image~=0.21.0
scipy~=1.10.1
opencv-python~=4.8.0.76
pytesseract~=0.3.10
pandas~=2.0.3
requests==2.31.0
torch~=2.0.1